refactor: create code structure with blueprints
- add services, blueprints, separate routes
- fix bug with the same diagram in all reports
