+++ .gitignore
@@ -1,3 +1,3 @@
 /static/*.mp4
 distracted_creds.json
-/__pycache__
+**/__pycache__
