+++ models/driver.py
@@ -1,0 +1,12 @@
+from flask_sqlalchemy import SQLAlchemy
+
+from extensions import db
+
+class Driver(db.Model):
+  __tablename__ = 'drivers'
+  id = db.Column(db.Integer, primary_key=True)
+  name = db.Column(db.String(80), nullable=False)
+  folder_id = db.Column(db.String(80), unique=True, nullable=False)
+
+  def __repr__(self) -> str:
+    return f'<Driver {self.name}'
