+++ models/user.py
@@ -1,14 +1,13 @@
-from flask_login import UserMixin, LoginManager
+# from flask_login import UserMixin, LoginManager
+from flask_login import LoginManager, UserMixin
 from flask_sqlalchemy import SQLAlchemy
 
-db = SQLAlchemy()
-login_manager = LoginManager()
+from extensions import db
 
 class User(UserMixin, db.Model):
   id = db.Column(db.Integer, primary_key=True)
   username = db.Column(db.String(80), unique=True, nullable=False)
   password = db.Column(db.String(120), nullable=False)
 
-@login_manager.user_loader
-def load_user(user_id):
-  return User.query.get(int(user_id))
+  def __repr__(self) -> str:
+    return f'<User {self.username}'
