+++ readme.md
@@ -1,0 +1,2 @@
+Running app:
+python -m flask --app app --debug run
