--- templates/video_page.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Real-Time Video Processing</title>
-</head>
-<body>
-    <h1>Real-Time Video Processing</h1>
-    <img src="{{ url_for('video_ml', video_name=video_name) }}" width="960" height="640">
-    <p id="prediction"></p>
-</body>
-</html>
